+++ OpenC-1.0/compiler/selfhost/native_runtime/openc_sh5_runtime.h
@@ -47,6 +47,7 @@ oc_text ocb_path_directory(oc_text value);
 
 uintptr_t ocb_process_argument_count(void);
 oc_text ocb_process_argument(uintptr_t index);
+oc_text ocb_process_executable_directory(void);
 oc_status ocb_process_run(
     oc_text command,
     int32_t *exit_code,